Search poi by location, type is Ok. Likeing is Ok.

+++ src/main/java/hu/andika/javaee/model/comment/CommentDao.java
@@ -16,6 +16,7 @@ public class CommentDao {
     protected String read;
     protected String readAll;
     protected String update;
+    protected String authorize;
     protected String delete;
 
     private static final DatabaseInit databaseInit = new DatabaseInit();
@@ -32,6 +33,7 @@ public class CommentDao {
         this.read = "SELECT * FROM comment WHERE id = ?";
         this.readAll = "SELECT * FROM comment";
         this.update = "UPDATE comment SET ";
+        this.authorize = "UPDATE comment SET isAuthorized = 'true' WHERE id = ?";
         this.delete = "DELETE FROM comment WHERE id = ?";
     }
 
@@ -106,6 +108,37 @@ public class CommentDao {
         return comments;
     }
 
+    public List<CommentDto> readAllCommentsToBeAuthorized() {
+        List<CommentDto> comments = null;
+        try (PreparedStatement pst = getConnection().prepareStatement(
+                "select c.id, c.content, c.userId, u.userName, c.pointOfInterestId, p.name, c.isAuthorized " +
+                        "from comment as c " +
+                        "left join user u on u.id = c.userId " +
+                        "left join poi p on c.pointOfInterestId = p.id " +
+                        "where c.isAuthorized = 'false'")) {
+            comments = new ArrayList<>();
+            ResultSet resultSet = pst.executeQuery();
+            while (resultSet.next()) {
+                comments.add(
+                    new CommentDto(
+                        resultSet.getInt("id"),
+                        resultSet.getString("content"),
+                        resultSet.getInt("userId"),
+                        resultSet.getString("userName"),
+                        resultSet.getInt("pointOfInterestId"),
+                        resultSet.getString("name"),
+                        Boolean.parseBoolean(resultSet.getString("isAuthorized"))
+                    )
+                );
+            }
+            logger.info("Found " + comments.size() + " Comments from database to be Authorized.");
+            return comments;
+        } catch (Exception e) {
+            logger.error("There is no Comments in the table!");
+        }
+        return comments;
+    }
+
         public List<Comment> readAllByUserId(Integer userId) {
         List<Comment> comments = null;
         try (PreparedStatement pst = getConnection().prepareStatement("SELECT * FROM comment where userId = ?")) {
@@ -122,12 +155,37 @@ public class CommentDao {
                         Boolean.parseBoolean(resultSet.getString("isAuthorized"))
                     )
                 );
-                System.out.println(resultSet.getString("content"));
             }
             logger.info("Found " + comments.size() + " Comments from database");
             return comments;
         } catch (Exception e) {
             logger.error("There is no Comments in the table with userID: " + userId);
+        }
+        return comments;
+    }
+
+    public List<Comment> readAllByPoiId(Integer poiId) {
+        List<Comment> comments = null;
+        try (PreparedStatement pst = getConnection().prepareStatement("SELECT * FROM comment where pointOfInterestId = ?")) {
+            pst.setInt(1, poiId);
+            comments = new ArrayList<>();
+            ResultSet resultSet = pst.executeQuery();
+            while (resultSet.next()) {
+                comments.add(
+                    new Comment(
+                        resultSet.getInt("id"),
+                        resultSet.getString("content"),
+                        resultSet.getInt("userId"),
+                        resultSet.getInt("pointOfInterestId"),
+                        Boolean.parseBoolean(resultSet.getString("isAuthorized"))
+                    )
+                );
+            }
+            logger.info("Found " + comments.size() + " Comments from database");
+            return comments;
+        } catch (Exception e) {
+            logger.error("Exception is: " + e);
+            logger.error("There is no Comments in the table with poiId: " + poiId);
         }
         return comments;
     }
@@ -165,6 +223,19 @@ public class CommentDao {
                 logger.error("SQL Delete failed!");
             }
             logger.info("The Comment with ID: " + id + " was successfully deleted!");
+        } catch (SQLException e) {
+            logger.catching(Level.ERROR, e);
+        }
+    }
+
+    public void authorizeById(Integer id) {
+        try (PreparedStatement pst = getConnection().prepareStatement(this.authorize)) {
+            pst.setInt(1, id);
+            int affectedRow = pst.executeUpdate();
+            if (affectedRow <= 0) {
+                logger.error("SQL Update failed!");
+            }
+            logger.info("The Comment successfully authorized!");
         } catch (SQLException e) {
             logger.catching(Level.ERROR, e);
         }
